Avoid wrap around on large numbers from requestBackupTime()
BackupManagerService can end up in an infinite loop if a
backup fails and is rescheduled a very long time in the
future.

The problem is if transport.requestBackupTime() returns a very
large value, for example "2^63 - currenttime", it can cause an
overflow turning the time when next try should be done into
a negative value. The negative value will trigger an alarm
directly instead of a long time in the future as intended,
totally the oposite than was intended.

This is corrected by checking that the number returned from
requestBackupTime() will not cause an alarm to be set at a
negative time. Instead the alarm is not set when overflow
occurs.

Change-Id: Id1f435e77e5e2c1e3407e26acf0efeb36ef56d81