Update patch set 6
Patch Set 6:

Hi Guillaume,

Yes, I agree that currently, webView.loadData() is not ideal and it is a shame that webView.loadData(buffer, "text/plain", "utf-8") no longer works as it used to. However, the fact that passing "utf-8" for the encoding argument used to work was not intentional. Doing so actually caused loadData() to generate an invalid URL, as the "charset=" part of the parameter was missing from the URL. So apps relying on this behaviour were using the API incorrectly (though I agree the vague documentation did not help) and thus the behaviour was not guaranteed. It is for this reason that the API was 'tightened up' in https://android.googlesource.com/platform/frameworks/base/+/f95d490b787f88dbd848a15f2a24b9ea929874db.

I don't think your proposed change is a good one. You're effectively combining the URL's encoding, and the parameter part of the mediatype, into a single argument to loadData(), which is confusing. Also, your code assumes that the first part of this argument is the value of a parameter named "charset". This is not a good assumption, as the set of valid parameter names depends upon the type/subtype of the mediatype. It is for this reason that I proposed using a "String mediaType" argument for the 'new' version loadData(), so as to keep the mediatype and its parameters together.

If you really think it's important to preserve the previous behaviour when passing "utf-8" for the encoding argument, I suggest you add code to special-case this, and generate a valid URL containing "charset=utf8".

Patch-set: 6
Label: Verified=0