Update patch set 1
Patch Set 1: Code-Review-1

I installed the font on my Nexus 10, and find that it _almost_ works but not quite. The problem is that the two tone accents, while they're in the correct order now, collide with each other. I made a screenshot:

http://levien.com/webfonts/i18n/DroidSansThai-change.png

I analyzed the font carefully, and am concerned that it makes a large number of changes that are hard to understand. This may be an artifact of using FontForge (the "FFTM" table is a pretty strong signal for this), which in my experience often makes potentially risky changes to a font on round-trip. Even the number of glyphs has changed, which should not be necessary for a change that only affects glyph positioning.

I'm going to pursue a change that uses GSUB reordering for the tone mark and the NIKHAHIT following, when that NIKHAHIT comes from a SARA AM. It may not be the best solution long-term, but feels satisfying to me in terms of making a minimal change least likely to break other things.

I also want to point out http://cgit.freedesktop.org/harfbuzz/tree/src/hb-ot-shape-complex-thai.cc#n270 , which suggests that it might have been premature to say that the font data is wrong - it does seem to render correctly with the Uniscribe renderer. Even though the tone mark reordering is not part of the public spec, it does seem to be part of the Uniscribe implementation. I think it's reasonable to try to make the rendering behavior of Harfbuzz match that of Uniscribe wherever possible (this will help a lot with web fonts and apk-provided fonts, as well as system fonts). In the meantime, however, I think a change to the font data is probably the best way to make it work for users.

Patch-set: 1
Reviewer: Gerrit User 1015625 <1015625@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1