Update patch set 1
Patch Set 1:

Yes, obviously, although I thought this would be a first step.

In https://android-review.googlesource.com/43372, I've got an initial CTS test that tries to test interpretation of the decoded data, which at the same time serves as an example on how to do it, something that is totally missing at the moment. That example uses these MediaFormat keys.

Patch-set: 1