Update patch set 1
Patch Set 1:

Hi Christopher,

We tried to fill /data partition both with enabled and disabled WAL. Phone was able to boot with WAL disabled. We checked that default DB journal mode was used instead of WAL.

Problem with WAL is that it uses "unreliable" disk space allocation (ftruncate). On some file systems (e.g. ext4) it is not guarantied that requested amount of disk space is available on partition after successful call to ftruncate(). As a result, when sqlite tries to touch allocated WAL buffer (which is mapped to memory using mmap), kernel sends SIGBUS signal the the application.
From user perspective we can observe random native crashes when /data is full.

To solve this situation the following should be done:
- Use "reliable" disk space allocation (subject of separate contribution). Allocation should fail when no disk space is available. 
- disable WAL in SettingsProvider. The reason is described in commit comment - SettingsProvider should always be able to open and read DB.

One more comment: when WAL is enabled, sqlite modifies DB-version in it's header to 2. During subsequent open calls, WAL will be enabled by default. That is why we manually do a call to db.disableWriteAheadLogging(), it this case sqlite switches back DB-version to 1 and WAL will not be applied automatically.

Patch-set: 1
Label: Code-Review=0