Fix misleading PendingIntent documentation.
The javadoc for PendingIntent.getActivity() and others says that requestCode
parameters is "currently not used". To me that implies that its value doesn't
affect anything, but in fact it does. It even has a useful function: to work
around the unexpected behaviour that different extras don't make intents
different. I added more of a hint about that too, since it is a common source
of surprise and frustration (as mentioned in the introduction).

Change-Id: Ica8d24068ee2ea5e5176b81a21682970ae6ca737