Sometimes routing to headset is not happening when BT is connected.
While BT is ON, when the wired headset is slowly inserted, the transistion from headphone
to headset is causing this issue. In this transistion, the headset gets connected first
and then the headphone gets disconnected. Since the last operation was disconnection the
output device changes to BT. This fix will check whether either of the device(headset or
headphone) is connected before changing the device to BT.

Change-Id: Ic75cc6a724b749512ddd9d7cdcddbb2cd8123fec
Signed-off-by: Mazhar <mazharx.sameullah@intel.com>