Update patch set 2
Patch Set 2: Code-Review-1

This isn't a good idea.

If the cursor was closed while the adapter was in use then the app should indeed crash.  If we allow this call to proceed normally then we'll have a known situation where an adapter is operating on a closed cursor.

Please try to find the root cause of the closed cursor and fix that.

Patch-set: 2
Reviewer: Gerrit User 1003957 <1003957@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1