Update patch set 5
Patch Set 5: Code-Review-1

I think this change is using the wrong kinds of defaults.

Fundamentally the information we're trying to communicate to the system is that it doesn't have a battery.  So we should just tell it that.

Why are you faking it to assume that there is a battery?  Instead I'd set "ac online" to true and "battery present" to false.

Also, if we've determined there is no power supply directory in sysfs, then we shouldn't bother trying to read from it or applying individual defaults to each sysfs file.

Just set something like "hasPowerSupply" to false then whenever it's false, skip reading all of the files and set the fields to match the assumption that the device is AC powered.

Patch-set: 5
Reviewer: Gerrit User 1003957 <1003957@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1