load BackupAgents with shared libraries
When loading BackupAgents, add shared libraries to the class path.

For example, let's suppose browser contains "A" library using <uses-library>.
During backup or after backup, if access browser, ClassNotFoundException
occurs on and off. Because "A" library is not linked, unexpectedly.
To fix it, add shared libraries to the class path as loading backup agent.

Change-Id: I1a51e485af2748c57cb6a29ed50a0246bb820af6