Update patch set 2
Patch Set 2: Code-Review+1

Jeff...
I guess that this is really a question of whether you wish to bullet-proof the server side or rely on the client not doing stupid things.

In this case, the flaw was found by a static analysis tool. It IS a flaw... but not one that a perfect program will find.

We would prefer to clean these sorts of things up where possible. It makes for more robust code.

-Bruce

Patch-set: 2
Reviewer: Gerrit User 1003360 <1003360@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1