Update patch set 1
Patch Set 1:

This fixes the bug mentioned below.

https://plus.google.com/118417777153109946393/posts/fryTTSjVvDA

It seems Nick did indeed file a bug for it but the change that fixed it for some of the drawables got reverted, probably because it changes the size of some icons as can more easily be seen here:

https://github.com/CyanogenMod/android_frameworks_base/commit/962daf9b1c4b12aff68a137472b6a7f4db123c65

This is the revert:

https://android.googlesource.com/platform/frameworks/base/+/f14f01cc26830459e4e4e0a601e5b5f810e0ca5f

Note that all these drawables were modified by me since I only learned about the reverted commit after committing it.

Patch-set: 1