Update patch set 1
Patch Set 1: Code-Review-1

This is an incorrect usage of ACTION_MEDIA_SCANNER_SCAN_FILE.  The Intent is supposed to be sent *to* the media scanner to request it to scan a specific file.  It is not supposed to be sent by the media scanner.

   /**
     * Broadcast Action:  Request the media scanner to scan a file and add it to the media database.
     * The path to the file is contained in the Intent.mData field.
     */
    @SdkConstant(SdkConstantType.BROADCAST_INTENT_ACTION)
    public static final String ACTION_MEDIA_SCANNER_SCAN_FILE = "android.intent.action.MEDIA_SCANNER_SCAN_FILE";

Patch-set: 1
Reviewer: Gerrit User 1003957 <1003957@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1