Center the tab's text whenever possible
The current implementation of a TabView always use the default 
horizontal gravity of TextView (i.e Gravity.START) to align the
text. While this is not an issue when having a single line text
(laid out using WRAP_CONTENT), it may be a problem when having a
two-line text (see https://dl.dropbox.com/u/1484156/tab_not_centered.png)

This patch consists on forcing the gravity to CENTER_HORIZONTAL
when the TabView is used as a tab (and not as an list item) and
the icon is not visible (i.e the tab contains only some text).

Change-Id: Ib81cdf29c9bbae53ebcdb2f866ecc1667da3b355