View: Adjust primary and secondary buttons behaviour
BUTTON_PRIMARY and BUTTON_SECONDARY are used by external mouses, touchpad controls and,
partially, by stylus. This patchset returns the expected behaviour for this buttons.
Now, both buttons perform the same as a finger touch (for click and long click).
When Android is controlled under a mouse or touchpad  the user experience should look like to
a normal pc experience:

- left click: simple click (no long press)
- rigth click: context menu (long press or context menu, but not a simple click)

Change-Id: I1abc7bf9275100e3df2b62fc93ba4d63a624f4ae
Signed-off-by: Jorge Ruesga <j.ruesga.criado@gmail.com>