Update patch set 1
Patch Set 1:

If we want to go this direction, AbsListView should cache the adapter's areAllItemsEnabled result in a field between data set changes/new adapters being set since an adapter's areAllItemsEnabled method is not guaranteed to run in constant time. (The documentation for the method technically states that the result should not change, but it's questionable how closely this advice has been followed in apps.) ListView already defines a field for this, mAreAllItemsSelectable, and only updates it when a new adapter is set. Perhaps this should be pushed down to AbsListView and renamed?

In any case, there are other possible bug-compatibility concerns with this patch as the result of this method is little-used and the widely extended BaseAdapter's default implementation always returns true.

Patch-set: 1
Reviewer: Gerrit User 1003263 <1003263@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0