Update patch set 1
Patch Set 1: Code-Review-1

We need this patch since there is already a way of specifying the min and max date by getting the DatePicker via DatePickerDialog.getDatePicker() which exposes APIs for setting the min and max date.

You need to change the layout files that declare DatePickerDialog's interface to specify the minDate and the maxDate attribute of the DatePicker contained in the dialog. This way the defaults will be reasonable while client code can change the min and the max. Feel free to prepare a patch :)

Search for date_picker_dialog.xml and to all files set the min and max date attributes. For example:

<DatePicker xmlns:android="http://schemas.android.com/apk/res/android"
    android:id="@+id/datePicker"
    android:layout_gravity="center_horizontal"
    android:layout_width="wrap_content"
    android:layout_height="wrap_content"
    android:spinnersShown="true"
    android:calendarViewShown="false"
    android:minDate="01/01/1902"   // Add this
    android:maxDate="12/31/2036"  // Add this
    />

Patch-set: 1
Reviewer: Gerrit User 1010257 <1010257@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1