Update patch set 2
Patch Set 2: Code-Review-1

The problem is in the DatePickerDialog not the DatePicker widget. The former is using DateUtils to format the dialog's title and the DateUtils class does not work with 
dates outside to the specified range.

The reason for the DateUtils class not being able to format dates outside the range of 1902 and 2016 is because internally it is using the Time class which does not support such dates.

We should avoid changing the defaults since some apps may rely in the fact that by default the DatePicker could choose dates before 1902 and after 2036.

The right fix is to update the DateUtils' format method to use Calendar instead of Time.

Patch-set: 2
Label: Code-Review=-1