frameworks/base: Make device to be in listen state if inviation result=1.
When p2p_invite persistent=<id> is used to request a persistent group to
be re-invoked, the peer may reply with status=1 (info not yet available)
if upper layer processing of the invitiation is requested. The peer is
ten expected to start another invitation exchanged within 120 seconds if
the user authorizes the connection.

For this mechanism to work, the device that starts the invitation needs
to start listen mode to be able to receive the invitation request from
the peer. So start a p2p_listen of 120 when status=1.

Change-Id: I87681d606a03296f1df64ccd0cd8dae0190094a6