Update patch set 2
Patch Set 2:

But still, Patch Set 2 works only in Activity based AppWidget hosts such as Homescreen, and won't work on the Keyguard, regarding the remote linkified textview with any clickable URLs. As the Keyguard cannot start an Activity w/o specifying NEW_TASK flag in the target intent. As a result, the system service crash&restart on any of the remote textview with URLSpan clicked. In terms of working with Keyguard AppWidget, the AppWidget Developer has to avoid any clickable URLSpans on a remote textview.

Patch-set: 2