Don't trigger empty draw in navigation bar deadzone
The nav bar registers for outside touch events. As a result it gets
ACTION_OUTSIDE events every time a touch down occurs anywhere else in
the UI. For each of these, DeadZone calls postInvalidate(), which triggers
a draw. However, the onDraw() method does nothing unless the dead zone is
configured to flash, which it normally isn't. These empty draws can delay
the foreground app's initial reaction to touch.

Modified to not call postInvalidate() when the draw will do nothing.

Change-Id: I828dc636d77ad0d8c810917e7347025106977372