Update patch set 1
Patch Set 1:

Hi,

You are absolutely right.

This fix is not directly for user's benefit but for passing strict stability tests often required by operators. OEMs should pass the strict tests such that "device should not crash even after 100 hours of video playback" and so-on. We have bunch of those fixes in many places and this is one of those. These things are not essential but could help to deliver Android devices to end users earlier :)

But if you think this isn't good, I'm fine to abandon it. Just it would be great if I could know your idea if you have time.

Thanks,
Aimi

Patch-set: 1
Reviewer: Gerrit User 1017042 <1017042@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0