Reset resolved Drawables on AttachedToWindow
Since unattached views (inflated in e.g. adapters) always resolves
the layoutDirection as LTR, the drawables should be resolved again
after onAttachedToWindow has been called.
It is also required that the resetResolvedDrawables method reset
the left and right drawable if start or end isn't null. This gives
precedence to Start and End over Left and Right.

Change-Id: I85323f102a1cc23d399e5d8dc29d87e59162a92c