Update patch set 1
Patch Set 1:

It seems that when the injector is destroyed the Java objects exposed to JavaScript are not removed. The destroy method nullifies that variable holding one of these objects. Now if JavaScript calls into Java we sync on a null instance. Generally using a nullifiable variable for sync is not a good idea.

Can you also fix the destroy() method which is now:

    public void destroy() {
        if (mTextToSpeech != null) {
            mTextToSpeech.shutdown();
            mTextToSpeech = null;
        }

        if (mCallback != null) {
            mCallback = null;
        }
    }

The right implementation is:

    public void destroy() {
        removeAccessibilityApisIfNecessary() 
    }

Patch-set: 1
Reviewer: Gerrit User 1010257 <1010257@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0