Update patch set 1
Patch Set 1:

looking at the code, I think some raciness or some error conditions (dispatchOnInit()) in calling addaccessibilityapisifnecessary could be causing this, so the change looks ok. 

I don't think I understand the suggested motivation to fix destroy() though. What is the problem and will the change really fix it? There is a comment above removeAccessibilityXX method that says it should only be called before a page load. I guess this is there because if we remove the api on destroy, we can break the executing JS script rather than gracefully shutting down. The injected inner class has its own mtexttospeech variable so not removing it will not cause a failure on destroy, I think.

Patch-set: 1