Update patch set 1
Patch Set 1: Code-Review-1

My suggestion was just as a good practice but yeah this will not solve the NPE issue. Actually this change is not correct since adding another object as a lock does not save you from the NPE exception a few lines below - see speak and stop methods for example.

Patch-set: 1
Label: Code-Review=-1