Update patch set 1
Patch Set 1:

Thanks for the review comments.

Actually the problem occurs only in the object creation of inner class TextToSpeechWrapper. The below is a problem:

mTextToSpeech = new TextToSpeech(
                    context, mInitListener, null, pkgName + ".**webview**", true);

The listener is notified instantly in case of error, which results in NPE (as locked on inner class mTextToSpeech which is still not created). To have a proper sync, I put a new object lock for all the methods.

The main class operations are being performed checking mTextToSpeech to null. so there is no chance of NPE in main class.

This patch avoids the NPE for the initlistener for the inner class only.

Patch-set: 1
Reviewer: Gerrit User 1013051 <1013051@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0