Add reporting of TLS errors, TLS alerts, and EAP notifications to wifi framework.
This patch creates a WifiManager.SUPPLICANT_INFO_ACTION broadcast intent that
can be used to pass information about the authentication to other apps that
may be able to use it to better determine why an authentication failed.  TLS
errors and EAP notifciations are message types already supported in existing
versions of wpa_supplicant.  The TLS alert message is a new one that is being
submitted upstream.  However, supporting the TLS alert message will have no
impact on functionality until an appropriately patched version of wpa_supplicant
is included.

Change-Id: I541e2f79055f3f4b461e307d24728f80e49367d4
Signed-off-by: Chris Hessing <chris.hessing@cloudpath.net>