Cached View should be identified by both name and class loader.
In Java, a class is identified by both name and class loader.
If a same class is loaded by different class loaders, casting
between such class instance will cause a ClassCaseException.

android.support.v4.view.ViewPager is not a pre-load class,
classloader of this class is PathClassLoader and dexpath is
the file path of apk. If 2 apks are run in a same process
and both APKs use ViewPager, it will cause ClassCaseException.

Change-Id: I49d412be1f43ff64c52513c07088692fde937df5
Signed-off-by: Bin Li <libin@marvell.com>