Update patch set 3
Patch Set 3:

Thanks for taking a look at this!

Patch set 3 adds the caps lock to backspace remap. I wasn't sure whether to put this in or not, as some Colemak layouts do not include it (notably those included with Mac OS X and iOS). But it's probably best to follow the colemak.com layout, so here it is.

Patch-set: 3
Reviewer: Gerrit User 1017758 <1017758@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0