Send volume state for emulated storage after boot
The state of the emulated storage is not broadcast
at phone startup. This can lead to appications ending
up in the wrong state. For example applications waiting
for MEDIA_MOUNTED before running will not work.

This behaviour has changed since older versions of the
platform.

The state of the emulated storage is now broadcast when
the system is ready.

Change-Id: I113f060dd207beed009e38e6dfdaa23df6bab71a