make wifiDisplay connection works when display turns off
  A side effect of wifi optimisation advanced option is to make
P2P connection fails when the display turn off at P2P connecting
time.
  In fact, some filters are enable at wifi driver level and broadcast
frame are no more received. If the DUT is Group Owner, the  DHCP
frames sent by the adapter (as broadcast) are reject and IP address
is no provided to the adapter and so connection fails.

  This fix prevent to enable SUPEND MODE at wifi driver level during
a P2P connection by handling the CONNECTING state. It also prevent to
sent legacy scan during the P2P connecting phase because the delay
introduced by this kind of scan could lead to a P2P group provisioning
timeout.

Signed-off-by: Honore Tricot <honorex.tricot@intel.com>
Signed-off-by: chenxi2x <xiaolongx.chen@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Change-Id: Ide056e91b78390fdf2139e15e20c6463982ec72f