add a flag control to enable the capability - set default locale from USIM;
this also fix a bug about system default locale, with JB 4.2.2, setLocale(Local Loc) function will
set userSetLocale as 'true'; when ActivityManagerService intialize locale on power up, it will make
the default locale as 'persist'. according to the design, only when user set the locale from UI, the
userSetLocale will be set as 'true', and the user setting is 'persist'.

Change-Id: I82d4340a04a25b2912f7a577682028335bf6740a