Reuse notification if order unchanged
When a notification was updated, the status bar assumed that its
order in the list changed if its time or priority changed. When
the assumption was wrong, the notification would appear to flicker
in place as its view was destroyed and re-created. Also, if the order
should change, the status bar assumed that notifications at the top of
the list did not need to move and so reused the view. Notifications
at the top of the list that changed to a lower priority would not move
down as a result.

Modified to determine order changes accurately, instead of relying
on those assumptions.

Change-Id: Ic4bec6be4df7427637fd6c7de6ebaf2206330cbb