Corrected mediaframeworktest test1_0ConstructorFromType
The test case queries for the first effect in a list and then try
to recreate it by using the UUID of that effect.
This will fail if the first effect in the list is not implemented as
a standard type and instead have the type set to
AudioEffect.EFFECT_TYPE_NULL, hence reconstruction of it will fail.
The change made in this commit will look into the list of effects and
find the first one that is NOT of type AudioEffect.EFFECT_TYPE_NULL.

Change-Id: Ic01cd77369c066dd8020ab9606525c2ca5c0d8ac