Update patch set 1
Patch Set 1:

This is an interesting model but it's not clear to me that it converges on a meaningful velocity vector for the finger itself.

I might be mistaken so I'll restate what I am reading in the code.

The algorithm assumes the initial velocity beyond the horizon is always 0.

Then for each sample in time, it computes the average velocity using distance traveled between oldest and most recent point.  (This is prone to underestimating velocity when a change of direction occurs.)

It does some magic to translate the velocity into a force vector and integrates acceleration.  I'm not clear on this part here because I can't think of a physical mechanism by which velocity would translate into force.  We could possibly say something about static vs. dynamic friction but it doesn't seem to fit here.

When we're done, we've estimated the velocity of a moving pane of glass acted upon by the finger.  So we've modeled the behavior of a ListView but the physics don't seem quite right to me.

The trouble is that we're actually trying to measure the velocity of the finger and we do want to be sensitive to relatively sudden changes in direction.

After all, your finger does not experience any significant friction on the glass when it changes direction.  It's velocity isn't really impeded in any way.

But we do want to filter out spurious changes and measurement errors...

Patch-set: 1
Reviewer: Gerrit User 1003957 <1003957@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0