Update patch set 1
Patch Set 1:

> The algorithm assumes the initial velocity beyond the horizon is always 0.

Yes.  This could be changed, although I haven't found it necessary.

> Then for each sample in time, it computes the average velocity using distance traveled between oldest and most recent point. (This is prone to underestimating velocity when a change of direction occurs.)

Not quite.  At each sample it computes a current finger velocity based on the distance and time between the previous and current point.

> It does some magic to translate the velocity into a force vector and integrates acceleration. I'm not clear on this part here because I can't think of a physical mechanism by which velocity would translate into force. We could possibly say something about static vs. dynamic friction but it doesn't seem to fit here.

It's akin to friction, but a closer physical mechanism would be viscous damping - imagine the surface has a coating of viscous fluid that the finger is moving through.  However, this is not intended to model a real physical system.  The key is that the force is proportional (with exponent) to the delta between the current finger and surface velocities.  This force is exerted on the surface, and the effect is that the surface velocity converges on the finger velocity.  The exponent determines the strength of this effect and therefore how rapidly it will converge.

> When we're done, we've estimated the velocity of a moving pane of glass acted upon by the finger. So we've modeled the behavior of a ListView but the physics don't seem quite right to me.

I chose it because it can converge rapidly on the finger velocity, but it does still have a damping/filtering effect which makes it tolerant to jitter and noise in the finger data.

> The trouble is that we're actually trying to measure the velocity of the finger and we do want to be sensitive to relatively sudden changes in direction.

The sensitivity is tunable with the exponents, ACCELERATION_EXP and DECELERATION_EXP.  I have it tuned to be more sensitive to acceleration and less sensitive to deceleration, so that a fling wouldn't be significantly slowed down by a small "hook" at the end of the gesture

> After all, your finger does not experience any significant friction on the glass when it changes direction. It's velocity isn't really impeded in any way.
But we do want to filter out spurious changes and measurement errors...

Think of it more as damping the surface movement than impeding the finger.  It allows a little "slip" between the surface and the finger in order to provide filtering.

Patch-set: 1