Update patch set 1
Patch Set 1: Code-Review-2

I dont think this is correct. 

From ZoomButtonsController.java:

    private FrameLayout createContainer() {
        LayoutParams lp = new LayoutParams(LayoutParams.WRAP_CONTENT, LayoutParams.WRAP_CONTENT);
        // Controls are positioned BOTTOM | CENTER with respect to the owner view.
        lp.gravity = Gravity.TOP | Gravity.START;
...

        FrameLayout container = new Container(mContext);
        container.setLayoutParams(lp);
...

        LayoutInflater inflater = (LayoutInflater) mContext
                .getSystemService(Context.LAYOUT_INFLATER_SERVICE);
        inflater.inflate(com.android.internal.R.layout.zoom_container, container);

        mControls = (ZoomControls) container.findViewById(com.android.internal.R.id.zoomControls);
...
}

It means that the Zoom controls are positioned at the start of the View (in LTR on the right and in RTL mode on the left). 

Then in RTL mode the position of the +/- buttons needs also to be layout mirrored (verified with a native Arabic speaker).

Patch-set: 1
Reviewer: Gerrit User 1005088 <1005088@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2