Update patch set 2
Patch Set 2: Code-Review+1

When a new Ad-Hoc network is configured (as opposed to connecting to an existing one) it needs to be configured for a specific channel. The User Interface (Settings) needs to show a list of possible available channels to let the user select an operating channel. Channel availability depends on regulatory rules and HW support, so it needs to be dynamic.

Patch-set: 2
Reviewer: Gerrit User 1020464 <1020464@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1