Fix for force finishing activity call is stuck in an infinite loop.
In a case where there is only Launcher activity in the activity stack and
a new activity WallpaperChooser from same Launcher2 package is started
through adb shell then Launcher activity is finished forcefully. It is
finished and started back. But it traps in an infinite loop being finished
and started continuously hence watchdog kills system process resulting in
framework reboot.

Change-Id: Iba23a62a029a004e0f3457f7fc538ce1f3e15677