Implement surface displays
Surface displays introduce a new powerful feature.  A new Display (TYPE_SURFACE)
is added, and can be created by any application or service (note that although
this patch keeps the API @hidden, no special permissions are required to use it.)
To create such a display, a Surface must be passed in, obtained from e.g. a
SurfaceTexture.  This Surface is handed to surfaceflinger to manage.  The
SurfaceTexture or other endpoint then receives updates from surfaceflinger
whenever that "physical" display is updated.  This enables many use cases, like
easily drawing and updating full view hierarchies into GL textures, or producing
an encoded video stream of an active view hierarchy and providing a privately
managed external display through some application-specific protocol.

In order to prevent security violations, TYPE_SURFACE displays are associated
with the PID and UID that created them.  They are visible and accessible only to
that UID, and when the PID that created them dies, they are automatically removed.
In addition, TYPE_SURFACE displays will never mirror another logical display.

In order to enable use cases which manage a private external display while the
managing service or app is not in the foreground, TYPE_SURFACE displays will never
be blanked, and so updates will continue even when the main screen is blanked.

There are three potential caveats with this patch that may need some additional
attention from Android display & frameworks people:

First, currently the only way to draw to a secondary Display is through a
Presentation.  Since Presentations are just special Dialogs, they get hidden
when their associated Activity does.  However, one of the primary use cases
for this feature is to manage a private external display within an app or service
which may be backgrounded while the main display remains interactive, and the
normal Presentation behavior makes this impossible.  The "right" solution to this,
perhaps, is to allow multiple Activity stacks to be simultaneously active and
associated with external displays.  This is quite a big change to the system,
however, much broader than the basic use case requires.  So, instead, I have
allowed (for TYPE_SURFACE displays only) Presentations to be created with the
application context instead of an Activity context.  This is a tiny change, but
it effectively does exactly what we want: it detaches the Presentation from any
Activity lifecycle and allows it to continue to function even when there is
no associated foreground Activity.  The limitation of this feature to TYPE_SURFACE
displays only prevents any contention or security issues.

Second, it appears there may be some insufficiency or bugs either in the current GL
fencing mechanism or in surfaceflinger's error handling, when testing on the Nexus 4.
Specifically, if a SurfaceTexture is created by an app and used to generate a Surface
for a TYPE_SURFACE Display, there is a potential for that app to kill surfaceflinger.
The condition that causes this is when there are other GL calls outstanding in the app's
GL context at the time when it calls SurfaceTexture.updateTexImage().  If this happens,
there are messages printed on logcat about a GL fencing timeout, and surfaceflinger
hangs, making the screen stop updating and requiring a reboot.  Adding a glFinish()
before the call to updateTexImage() works around this problem.  It would likely be
sufficient to add a call to glFinish() in the implementation of updateTexImage() to
avoid this specific issue, but this points to a possibly broader stability issue with
allowing uncontrolled endpoints for surfaceflinger output surfaces.  It would be good
to get an opinion from Mathias or other surfaceflinger gurus on this problem.

Finally, this is a self-contained patch that provides some useful functionality, but
it does not provide everything needed in order to enable private external displays.
For this, input must also be wired up, and routed properly to the UI on that display.
This requires some changes to the input architecture and some additional services for
event injection, and these are somewhat orthogonal to these display features, so a patch
enabling this will be provided in a separate commit.