Refactoring makeEntryFor() in MediaScanner
In makeEntryFor(), "LIKE" in SQL is used to get file entry
when case insensitive path option is enabled.
But, if path contains '_' OR '%' then "LIKE" regards this as literals.
It makes code more complicated and query is called twice at worst case.

"COLLATE" can replace the "LIKE".
In case of using "COLLATE nocase", it uses "path_index_lower" index.

Change-Id: I4eccd8b8cf24b12e801b98fa80430057aff6cc92