P2P: Fix for connection between GO and the 3rd client.
When GO sends an invite to p2p device, the p2p client may not
have the updated scan entries due to which the framework may
send p2p_connect without join which would result in a connection
failure. To address this, always set join true when invitation is
received at the client side.

Change-Id: I64e515457f714980350fbfce6fa5de25cdf4c42c