Update patch set 2
Patch Set 2:

Great. If "Wi-Fi Display" should be kept untranslated or if it should be translated to Wi-Fi-skärm depends on what the intended meaning was of the original English string.

It could either refer to the Wi-Fi Display standard/specification from the Wi-Fi Allience (http://www.wi-fi.org/knowledge-center/faq/what-difference-between-miracast-and-wi-fi-display), in which case it should probably not be translated.

It could also refer simply to a Wi-Fi-connected display, in which case the translation to Wi-Fi-skärm makes sense.

This might be a case where it is a good idea to ask the developers in the Android team what the intended interpretation is. Still, asking the linguist can't hurt either.

Patch-set: 2